tuples

+++ vllm/v1/sample/sampler.py
@@ -141,7 +141,7 @@ class Sampler(nn.Module):
 
         # Concatenate with the token_ids
         token_ids = token_ids.unsqueeze(-1)
-        sampled_logprobs = logprobs.gather(-1, token_ids)
+        sampled_logprobs = logprobs.gather(-1, token_ids.to(torch.int64))
         topk_indices = torch.cat([token_ids, topk_indices], dim=1)
         topk_logprobs = torch.cat([sampled_logprobs, topk_logprobs], dim=1)
